Splitting different polling logic into separate file. Split open group authentication into separate file and added header tests.
